fix(strategy): address CodeRabbit round-1 — single-source fee rule, shared factories, hardened tests
- market-gate: transferFeeRejectionReason is the single source for the
  transfer-fee rule (marketLegPasses delegates; the loop uses it for the
  reason string) — no duplicated copies to drift.
- dispatch-deps builders (runnerDispatchDeps / harvestDispatchDeps) hoisted
  to cycle scope, shared by the in-slot tail + idle-redeploy pass.
- evaluateHarvestGate takes a narrow HarvestGateConfig (no AppConfig cast).
- transfer-fee screen: plain typed leg entries (no cast / no string union).
- Test factories (makeTestLayer/makeAdapter/makeDatapiStats/DecisionRow/
  RecordedMemory) moved from decision-loop-exit.test.ts to bench/helpers.ts —
  harvest-gate imports them without re-registering the decision-loop suite.
- token-breaker race budget 1.5s -> 2.5s; harvest-gate asserts the claimFees
  spy (0 on skip, exactly 1 on healthy via a re-arming interval);
  transfer-fee tests cover malformed parsed data + the token-Y leg.
